Use boolean return values in native code instead of integers

+++ src/TorchSharp/PInvoke/LibTorchSharp.THSTensor.cs
@@ -247,7 +247,7 @@ namespace TorchSharp.PInvoke
         internal static extern long THSTensor_numel(IntPtr handle);
 
         [DllImport("LibTorchSharp")]
-        internal static extern long THSTensor_is_leaf(IntPtr handle);
+        internal static extern byte THSTensor_is_leaf(IntPtr handle);
 
         [DllImport("LibTorchSharp")]
         internal static extern IntPtr THSTensor_alias(IntPtr handle);
@@ -386,13 +386,13 @@ namespace TorchSharp.PInvoke
         internal static extern long THSTensor_copy_(IntPtr handle, IntPtr source, [MarshalAs(UnmanagedType.U1)] bool non_blocking);
 
         [DllImport("LibTorchSharp")]
-        internal static extern int THSTensor_is_contiguous(IntPtr handle);
+        internal static extern byte THSTensor_is_contiguous(IntPtr handle);
 
         [DllImport("LibTorchSharp")]
         internal static extern IntPtr THSTensor_contiguous(IntPtr handle);
 
         [DllImport("LibTorchSharp")]
-        internal static extern long THSTensor_is_pinned(IntPtr handle);
+        internal static extern byte THSTensor_is_pinned(IntPtr handle);
 
         [DllImport("LibTorchSharp")]
         internal static extern IntPtr THSTensor_pin_memory(IntPtr handle);
@@ -758,7 +758,7 @@ namespace TorchSharp.PInvoke
         internal static extern IntPtr THSTensor_isnan(IntPtr tensor);
 
         [DllImport("LibTorchSharp")]
-        internal static extern long THSTensor_is_nonzero(IntPtr handle);
+        internal static extern byte THSTensor_is_nonzero(IntPtr handle);
 
         [DllImport("LibTorchSharp")]
         internal static extern IntPtr THSTensor_isreal(IntPtr tensor);
